+++ config/php-wake-on-lan.php
@@ -5,7 +5,14 @@
  */
 return [
 
+    /**
+     * String containing target broadcast address in XXX.XXX.XXX.255 ipv4 format
+     */
     'broadcast_address' => '255.255.255.255',
 
-    'route' => '/wake-on-lan/{macAddress}',
+    /**
+     * Integer containing target port to use. Must be 7 or 9.
+     */
+    'port'              => 7,
+
 ];