+++ tests/MagicPacketTest.php
@@ -15,7 +15,7 @@ class MagicPacketTest extends TestCase
     /**
      * @covers \Diegonz\PHPWakeOnLan\MagicPacket::isMacAddressValid()
      */
-    public function testIsMacAddressValid()
+    public function testIsMacAddressValid(): void
     {
         $this->assertTrue(MagicPacket::isMacAddressValid('00:1B:21:1C:7F:23'));
         $this->assertFalse(MagicPacket::isMacAddressValid(' 00:1P:21:1X:7F:23'));
@@ -25,7 +25,7 @@ class MagicPacketTest extends TestCase
     /**
      * @covers \Diegonz\PHPWakeOnLan\MagicPacket::trimMacAddress()
      */
-    public function testTrimMacAddress()
+    public function testTrimMacAddress(): void
     {
         $this->assertEquals('001B211C7F23', MagicPacket::trimMacAddress(' 00:1B:21:1C:7F:23'));
     }
@@ -33,7 +33,7 @@ class MagicPacketTest extends TestCase
     /**
      * @covers \Diegonz\PHPWakeOnLan\MagicPacket::packMacAddress()
      */
-    public function testPackMacAddress()
+    public function testPackMacAddress(): void
     {
         $mac1 = '00:1B:21:1C:8F:23  ';
         $mac2 = '00:1C:21:1C:8F:27  ';
@@ -47,12 +47,12 @@ class MagicPacketTest extends TestCase
      * @covers \Diegonz\PHPWakeOnLan\MagicPacket::buildMagicPacketString()
      * @throws \Exception
      */
-    public function testBuildMagicPacketString()
+    public function testBuildMagicPacketString(): void
     {
         $mac1 = '00:1B:21:1C:8F:23';
         $mac2 = '00:1C:21:1C:8F:27';
-        $sample = \str_repeat(\chr(0xff), 6)
-                  .\str_repeat(pack('H12', '001B211C8F23'), 16);
+        $sample = str_repeat(chr(0xff), 6)
+                  .str_repeat(pack('H12', '001B211C8F23'), 16);
 
         $magicPacket1 = new MagicPacket($mac1);
         $magicPacket2 = new MagicPacket($mac2);