+++ tests/Socket/SocketTest.php
@@ -15,10 +15,10 @@ class SocketTest extends TestCase
     /**
      * @covers \Diegonz\PHPWakeOnLan\Socket\Socket::send()
      */
-    public function testSend()
+    public function testSend(): void
     {
         $socket = new Socket(SOL_UDP);
-        \socket_set_option($socket->getSocket(), SOL_SOCKET, SO_BROADCAST, true);
+        socket_set_option($socket->getSocket(), SOL_SOCKET, SO_BROADCAST, true);
         $result = $socket->send('', '255.255.255.255', 7);
         $this->assertEquals(0, $result, 'Bytes sent must be 0');
     }